remove (#29) : mapper 클래스 global 패키지로 이동
